+++ src/test/e2e_test.ts
@@ -209,6 +209,32 @@ suite('e2e', function() {
              assert.isAbove(ciAverage(diffBA.relative), 0);
            }));
 
+      test('multiple measurements', hideOutput(async function() {
+             const delayA = 20;
+             const delayB = 60;
+
+             const argv = [
+               `--config=${path.join(testData, 'multiple-measurements.json')}`,
+             ];
+
+             const actual = await main(argv);
+             assert.isDefined(actual);
+             assert.lengthOf(actual!, 2);
+             const [a, b] = actual!;
+             const diffAB = a.differences[1]!;
+             const diffBA = b.differences[0]!;
+
+             // We can't be very precise with expectations here, since
+             // setTimeout can be quite variable on a resource starved machine
+             // (e.g. some of our CI builds).
+             assert.isAtLeast(a.stats.mean, delayA);
+             assert.isAtLeast(b.stats.mean, delayB);
+             assert.isBelow(ciAverage(diffAB.absolute), 0);
+             assert.isAbove(ciAverage(diffBA.absolute), 0);
+             assert.isBelow(ciAverage(diffAB.relative), 0);
+             assert.isAbove(ciAverage(diffBA.relative), 0);
+           }));
+
       // Only Chrome supports FCP and CPU throttling.
       if (browser.startsWith('chrome')) {
         test('fcp', hideOutput(async function() {